Move everything .co packages
